Re-render on creation, loading animation
