+++ src/components/Classroom/TeacherClassroom.js
@@ -27,6 +27,7 @@ import TeacherNav from "../Navbar/teacher/teacherNav";
 import Assignments from "../Tabs/teacher/assignments/Assignments";
 
 import "../../pages/Timetable/Timetable.css";
+import "./NavTab.css";
 
 const TeacherClassroom = ({ dispatch, tt, classrooms, params }) => {
     const { user } = useSelector((state) => ({ ...state }));
